requirements.txt
discord.py>=2.3.0
colorlog>=6.7.0
yt-dlp>=2023.7.6
aiohttp>=3.8.0
jinja2>=3.0.0
markdown>=3.4.0
python-dotenv>=0.19.0
pynacl
dotenv